+++ cli/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added `--no-debug-syms` option to `inspect dump elf` sub-command
+- Added `--kallsyms` and `--kernel-image` options to `symbolize-kernel
+  sub-command
 
 
 0.1.8